+++ packages/nibijs/src/chain/chain.ts
@@ -137,21 +137,21 @@ export async function queryChainIdWithRest(
   return [chainId, err]
 }
 
-export async function isRestEndptLive(network: Chain): Promise<boolean> {
+export async function isRestEndptLive(chain: Chain): Promise<boolean> {
   let isLive: boolean = false
-  const [chainId, err] = await queryChainIdWithRest(network)
+  const [chainId, err] = await queryChainIdWithRest(chain)
   if (chainId) {
     isLive = true
   }
   return isLive
 }
 
-export async function isRestEndptValid(network: Chain): Promise<boolean> {
+export async function isRestEndptValid(chain: Chain): Promise<boolean> {
   let isLive: boolean = true
-  const [chainId, err] = await queryChainIdWithRest(network)
+  const [chainId, err] = await queryChainIdWithRest(chain)
   if (!chainId) {
     isLive = false
-  } else if (chainId !== network.chainId) {
+  } else if (chainId !== chain.chainId) {
     isLive = false
   }
   return isLive